Fixed location service attachment and detachment flow with correct FlutterEngine
There are 2 Flutter Engines being created in the app one is our SharedEngine and the other is created by Firebase for background notification handling.

The location service was getting assigned and updated based on the sequence of Flutter engine creation hence causing issues like sometimes it gets initialised and terminated properly and sometimes not.

Now it should be connected with only one engine (our SharedEngine).